\name{atomprop}
\alias{atomprop}
\docType{data}
\title{
Standard atomic weights
}
\description{
Data frame with atom names, symbols and standard atomic weights.
}
\usage{data(atomprop)}
\format{
  The format is a data frame with 117 rows and 4 columns.
}
\source{
From: http://iupac.org/publications/pac/78/11/2051/
}
\references{
Pure Appl. Chem., 2006, Vol. 78, No. 11, pp. 2051-2066
}
\examples{
data(atomprop)
atomprop[1:4,]
}
\keyword{datasets}
