\name{pubchemFPencoding}
\alias{pubchemFPencoding}
\docType{data}
\title{
Enncoding of PubChem Fingerprints
}
\description{
Data frame with bit positions and substructure specifications.
}
\usage{data(pubchemFPencoding)}
\format{
  The format is a data frame with 881 rows and 2 columns.
}
\source{
From: ftp://ftp.ncbi.nih.gov/pubchem/specifications/pubchem_fingerprints.txt
}
\references{
See: ftp://ftp.ncbi.nih.gov/pubchem/specifications/pubchem_fingerprints.txt
}
\examples{
data(pubchemFPencoding)
pubchemFPencoding[1:4,]
}
\keyword{datasets}
