\name{read.AP}
\alias{read.AP}
\title{
Read Atom Pair String from File into APset
}
\description{
Helper function to read atom pair (AP) string representations from tabular files that were generated by \code{sdfStream} while passing on \code{sdf2ap} with argument \code{type="character"}.
}
\usage{
read.AP(file = "matrix.xls", colid = "AP")
}
\arguments{
  \item{file}{
name of file to import; usually a file generated by \code{sdfStream} function with atom pair \code{character} output 
}
  \item{colid}{
name of column containing atom pair character strings
}
}
\details{
...
}
\value{
object of class \code{APset}
}
\references{
...
}
\author{
Thomas Girke
}
\seealso{
\code{sdf2ap}, \code{sdfStream}
}
\examples{
## Load sample data
library(ChemmineR)
data(sdfsample); sdfset <- sdfsample
\dontrun{write.SDF(sdfset, "test.sdf")

## Define descriptor set in a simple function
desc <- function(sdfset) {
        cbind(SDFID=sdfid(sdfset), 
              # datablock2ma(datablocklist=datablock(sdfset)), 
              MW=MW(sdfset), 
              groups(sdfset), 
              AP=sdf2ap(sdfset, type="character"),
              rings(sdfset, type="count", upper=6, arom=TRUE)
        )
}

## Run sdfStream with desc function and write results to a file called 'matrix.xls'
sdfStream(input="test.sdf", output="matrix.xls", fct=desc, Nlines=1000)

## Select molecules from SD File using line index from sdfStream
indexDF <- read.delim("matrix.xls", row.names=1)[,1:4]
indexDFsub <- indexDF[indexDF$MW < 400, ] # Selects molecules with MW < 400
sdfset <- read.SDFindex(file="test.sdf", index=indexDFsub, type="SDFset")

## Write result directly to SD file without storing larger numbers of molecules in memory
read.SDFindex(file="test.sdf", index=indexDFsub, type="file", outfile="sub.sdf")

## Read atom pair string representation from file into APset
apset <- read.AP(file="matrix.xls", colid="AP")
cid(apsdf) <- as.character(indexDF$SDFID)  
}
}
\keyword{ utilities }
